Broke out the utilities classes and methods into five new files; much more readable now.
